NOTICE - APP DISCONTINUED, NO LONGER MAINTAINED, LOCATION OF THE REPO CHANGED!
This app/project is no longer under use and being continued by the author. There will be no further fixes and updates even regarding the present bugs in the app. The location of the repo from github is also being transferred to another location mentioned in the NOTICE section of the README.md file